requirements.txt
Flask==1.1.2
pandas==1.2.1
tables==3.6.1
scikit-learn==0.24.1
OpenCC==1.1.1
jieba==0.42.1
google-cloud-logging
